Add scenario testing to Pineapple and export hof utility
