+++ experimental/scenario.test.js.psnap
@@ -1,0 +1,14 @@
+{
+  "Example({ num: 5 }) [4743a1ffd62f16e2287ec060fc6abe7a067ea2dfefcffa957177fd1558f601e6]": {
+    "value": {
+      "pineapples": 4
+    },
+    "async": true
+  },
+  "SecondExample({ num: 5 }) [4743a1ffd62f16e2287ec060fc6abe7a067ea2dfefcffa957177fd1558f601e6]": {
+    "value": {
+      "pineapples": 2
+    },
+    "async": true
+  }
+}
