+++ hash.d.ts
@@ -1,0 +1,6 @@
+/**
+ * @test 'Hello World'
+ * @test 'Kevin'
+ * @param {string} str
+ */
+export function hash(str: string): any;
