+++ index.d.ts
@@ -1,1 +1,3 @@
-export {};
+import Scenario from "./experimental/scenario.js";
+import { hof } from "./run.js";
+export { Scenario, hof };
