+++ methods.d.ts
@@ -1,0 +1,4 @@
+export function createArbitrary(method: any): (data: any) => any;
+export default engine;
+declare const engine: AsyncLogicEngine;
+import { AsyncLogicEngine } from "json-logic-engine";
