+++ symbols.d.ts
@@ -1,0 +1,2 @@
+export const SpecialHoF: "___SpecialHoF___";
+export const ConstantFunc: "___ConstantFunc___";
